==> Code Files/1. Website Extraction.ipynb <==
import numpy as np
import pandas as pd
pd.options.mode.chained_assignment = None  # default='warn'

df=pd.read_csv('C:/Users/MyPC/OneDrive/Documents/Inbiot VS/AICTE Combined List 1.csv', skiprows=[0],header=None)

df.head()

df.isnull().sum()

df= df.dropna(axis=0)

df.isnull().sum()

from googlesearch import search

df["WEBSITE"]=np.nan

df.head()

for i in range(0,df.shape[0]):
    query=df.iloc[i,1]
    for j in search(query, num=1, stop=1, pause=1):
        df.iloc[i,6]=j
        print(df.iloc[i,1],'\t',j)

df.columns=['AICTE ID','NAME','ADDRESS','DISTRICT','INSTITUTE','STATE','WEBSITE']

df.to_excel('C:/Users/MyPC/OneDrive/Documents/Inbiot VS/First 500/AICTE Combined List FINAL.xlsx')
