==> Code Files/2. Email Extraction try 1.ipynb <==
import pandas as pd
import re
from matplotlib import pyplot as plt
import urllib.request
import numpy as np
df=pd.read_csv('C:/Users/MyPC/OneDrive/Documents/Inbiot VS/FINAL/Final after website/AICTE Combined List FINAL.csv')

df.head()

df['EMAIL']=np.nan

df.head()

df.shape

for i in range(0,df.shape[0]):
    try:
        weburl=urllib.request.urlopen(df.iloc[i,7])
        data=weburl.read().decode('utf-8')    
        valid=r'\b[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+\.[A-Z|a-z]{2,}\b'
        email=re.search(valid,data)
        if email!=None:
            a=email.group(0)
            df.iloc[i,8]=a
            print(i,df.iloc[i,2],'\t',email.group(0))
        else:
            print(i,'\t', "Match not found")    
    except Exception as error:
        print(error)

df.to_csv('D:/Inbiot VS/EMAILS/Upto 5300/AICTE Combined List FINAL.csv')
