==> Code Files/4. Email Extraction try 3.ipynb <==
print("Kushal")

import pandas as pd
import urllib.request
from googlesearch import search
import re
df=pd.read_csv('C:/Users/MyPC/OneDrive/Documents/Inbiot VS/FINAL/Final after 2nd try/AICTE Combined List FINAL(2ndtryEmails).csv')

df.head(4)

df.shape[0]

# Trying with ISO-8859-1 decoding insttead of UTF-8

for i in range(0,df.shape[0]):
    if pd.isnull(df.iloc[i,9]):
        try:
            query=df.iloc[i,3]+"Contact Us"
            for j in search(query, num=1, stop=1, pause=1):
                weburl=urllib.request.urlopen(j)
                data=weburl.read().decode('ISO-8859-1')    
                valid=r'\b[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+\.[A-Z|a-z]{2,}\b'
                email=re.search(valid,data)
                if email!=None:
                    a=email.group(0)
                    df.iloc[i,9]=a
                    print(i,'\t',df.iloc[i,3],'\t',email.group(0))
                else:
                    print("Match Not found")
        except Exception as error:
            print(error)

df.to_csv('C:/Users/MyPC/OneDrive/Documents/Inbiot VS/Email 3rd try/Final/AICTE Combined List FINAL.csv')

df.isnull().sum()

# Trying with Headers defined(for HTTP error code 403:Forbidden)

header= {'User-Agent': 'Mozilla/5.0 (X11; Linux x86_64) ' 
      'AppleWebKit/537.11 (KHTML, like Gecko) '
      'Chrome/23.0.1271.64 Safari/537.11',
      'Accept': 'text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8',
      'Accept-Charset': 'ISO-8859-1,utf-8;q=0.7,*;q=0.3',
      'Accept-Encoding': 'none',
      'Accept-Language': 'en-US,en;q=0.8',
      'Connection': 'keep-alive'}

for i in range(0,df.shape[0]):
    if pd.isnull(df.iloc[i,9]):
        try:
            query=df.iloc[i,3]+"Contact Us"
            for j in search(query, num=1, stop=1, pause=1):
                req = urllib.request.Request(url=j, headers=header)  
                data=urllib.request.urlopen(req).read().decode('ISO-8859-1')    
                valid=r'\b[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+\.[A-Z|a-z]{2,}\b'
                email=re.search(valid,data)
                if email!=None:
                    a=email.group(0)
                    df.iloc[i,9]=a
                    print(i,'\t',df.iloc[i,3],'\t',email.group(0))
                else:
                    print(i,"Match Not found")
        except Exception as error:
            print(i," ",error)

df.isnull().sum()

df.to_csv('C:/Users/MyPC/OneDrive/Documents/Inbiot VS/Email 3rd try/Upto 4000 (headers)/AICTE Combined List FINAL.csv')

sub='.pdf'
df["a"]=((df["EMAIL"].str.find(sub))!=-1 )

import pandas as pd
df1=pd.read_csv('C:/Users/MyPC/OneDrive/Documents/Inbiot VS/Presentation/FINAL AFTER HEADERS/AICTE Combined List FINAL.csv')

df1.head()

df1.isnull().sum()
