==> Code Files/5. Website Occurences.ipynb <==
import pandas as pd
df=pd.read_csv('C:/Users/MyPC/OneDrive/Documents/Inbiot VS/Presentation/FINAL AFTER HEADERS/AICTE Combined List FINAL.csv')
import numpy as np

# Counting website Occurences based on netloc value
# (varities of website)

from urllib.parse import urlparse

df.shape

df.head(2)

ds = []
for i in range(0,df.shape[0]):
    string=df.iloc[i,6]
    o= urlparse(str(string))
    ds.append(o.netloc)

from collections import Counter
#Counter(ds).most_common()
a=Counter(ds).most_common()

#List-to-Dataframe
tempdf=pd.DataFrame(a) 

#renamed the columns
tempdf.columns=['WEBSITE','OCCURENCES']

## Following will be useful in separating 3rd party websites

tempdf

#tempdf.to_csv('C:/Users/MyPC/OneDrive/Documents/Inbiot VS/Occurences/websiteoccur.csv')

# 2nd Approach for counting occurence (actual occurences)

a=df['WEBSITE'].value_counts()

## Useful in Separating duplicates

a

#a.to_csv('C:/Users/MyPC/OneDrive/Documents/Inbiot VS/Occurences/Website 2nd approach/websiteoccur2.csv')
