==> Code Files/7. Dealing with Duplicate Websites.ipynb <==
import pandas as pd
origdf=pd.read_csv('C:/Users/MyPC/OneDrive/Documents/Inbiot VS/Filtering/No 3P Dupl Remove/NO_3rdparty.csv')
origdf.shape[0]

# Tried using drop_duplicates

origdf.drop_duplicates(subset=['WEBSITE','EMAIL'],inplace=True)
#origdf.drop_duplicates(subset=['WEBSITE'],inplace=True)

#origdf.to_csv('C:/Users/MyPC/OneDrive/Documents/Inbiot VS/Filtering/No 3P Dupl Remove/CLEANED/Experiment/duplremoved.csv')

# Still some duplicate website exist

a=origdf.WEBSITE.value_counts()
stilldup=(a[a>1]).to_frame()
stilldup.columns=['OCCURENECES']
stilldup

# BEST WAY IS TO SEPARATE AND REMOVE ALL ROWS WITH DUPLICATE

### (Note: I got the occurences for present file(No 3rdparty) using same code for website occurences)

df=pd.read_csv('C:/Users/MyPC/OneDrive/Documents/Inbiot VS/Occurences/Occurences after removing 3rd party/2nd approach/websiteoccur2.csv')

df

## Making a list of all websites with more than 1 occurences

search=list(df.WEBSITE[df.OCCURENCES>1])

search

origdf=pd.read_csv('C:/Users/MyPC/OneDrive/Documents/Inbiot VS/Filtering/No 3P Dupl Remove/NO_3rdparty.csv')
origdf.shape[0]

# Separating all websites which have multiple occurences
## (search, save and delete)

1. Search

b=origdf[origdf['WEBSITE'].str.contains('|'.join(search),na=False)]
b
#p=(b.index).tolist()
#df.drop(labels=p,inplace=True)

2. Save

#b.to_csv('C:/Users/MyPC/OneDrive/Documents/Inbiot VS/Filtering/No 3P Dupl Remove/Removed DUPLS/DUPLICATES.csv')

3. Delete

b=origdf[origdf['WEBSITE'].str.contains('|'.join(search),na=False)]
b
p=(b.index).tolist()
origdf.drop(labels=p,inplace=True)

## Exporting dataset with filtered websites

origdf.shape[0]

#origdf.to_csv('C:/Users/MyPC/OneDrive/Documents/Inbiot VS/Filtering/No 3P Dupl Remove/ACTUAL Cleaned/duplremoved.csv')
