==> Code Files/8. Dealing with DuplEmails.ipynb <==
import pandas as pd
df=pd.read_csv('C:/Users/MyPC/OneDrive/Documents/Inbiot VS/Filtering/No 3P Dupl Remove/ACTUAL Cleaned/Duplicates Removed  (No3party).csv',index_col=0)

df.head()

df.EMAIL.value_counts()

a=df.EMAIL.value_counts()
#a.to_csv('C:/Users/MyPC/OneDrive/Documents/Inbiot VS/Filtering/No 3P Dupl Remove/ACTUAL Cleaned/Email Occurences after duplremov/emailoccur.csv')

edf=pd.read_csv('C:/Users/MyPC/OneDrive/Documents/Inbiot VS/Filtering/No 3P Dupl Remove/ACTUAL Cleaned/Email Occurences after duplremov/emailoccur.csv')

edf

edf.columns=['EMAIL','OCCURENCES']

edf.head(3)

search=list(edf.EMAIL[edf.OCCURENCES>1])
search+=['.jpg','.png']

#search

# Separating Duplicate and Null Emails 
## (search, save and delete)

## 1. Searching Duplicate emails

b=df[df['EMAIL'].str.contains('|'.join(search),na=False)]
b.head(4)
#p=(b.index).tolist()
#origdf.drop(labels=p,inplace=True)

## 2. Saving Duplicate emails 

#b.to_csv('C:/Users/MyPC/OneDrive/Documents/Inbiot VS/Filtering/No 3P Dupl Remove/ACTUAL Cleaned/Duplicate and Useless Emails/Duplicate_and_useless_emails.csv')

## 3. Searching Null Emails

df.isnull().sum()

c=df[pd.isnull(df.EMAIL)]

c.head(2)

## 4. Saving Null emails

#c.to_csv('C:/Users/MyPC/OneDrive/Documents/Inbiot VS/Filtering/No 3P Dupl Remove/ACTUAL Cleaned/Null Emails/Null_emails.csv')

## 5. Deleting Duplicates Emails

b=df[df['EMAIL'].str.contains('|'.join(search),na=False)]
b.head(4)
p=(b.index).tolist()
df.drop(labels=p,inplace=True)

## 6. Deleting Null Emails

df = df.dropna(axis=0, how='any')

df.shape[0]

# FINAL Data for Zonal Distribution

#df.to_csv('C:/Users/MyPC/OneDrive/Documents/Inbiot VS/Filtering/No 3P Dupl Remove/ACTUAL Cleaned/FINAL FOR ZONAL DISTRUBTION/Final.csv')
