==> Code Files/9. Zonal Division.ipynb <==
import pandas as pd
df=pd.read_csv('C:/Users/MyPC/OneDrive/Documents/Inbiot VS/Filtering/No 3P Dupl Remove/ACTUAL Cleaned/FINAL FOR ZONAL DISTRUBTION/Final.csv')
import numpy as np

df.head(2)

# Northern Zone Colleges

search=["Himachal pradesh", "Punjab", "uttarakhand","Haryana","uttar pradesh"]
north=df[df['STATE'].str.contains('|'.join(search),na=False)]
north

#north.to_csv('C:/Users/MyPC/OneDrive/Documents/Inbiot VS/Zonal Distribution/NorthZone.csv')

# Southern Zone Colleges

search=["Andhra pradesh", "karnataka", "kerala", "Tamil nadu"]
south=df[df['STATE'].str.contains('|'.join(search),na=False)]
south

#south.to_csv('C:/Users/MyPC/OneDrive/Documents/Inbiot VS/Zonal Distribution/SouthZone.csv')

# Eastern Zone Colleges

search=["Bihar", "Sikkim", "Arunachal", "Nagaland", "Manipur","Tripura","Meghalaya","Assam","WB","Jharkhand","Odisha","Chattishgarh"]
east=df[df['STATE'].str.contains('|'.join(search),na=False)]
east

#east.to_csv('C:/Users/MyPC/OneDrive/Documents/Inbiot VS/Zonal Distribution/EasternZone.csv')

# Western Zone Colleges

search=["Rajasthan","Madhya Pradesh","Gujarat","Maharashtra","Goa"]
west=df[df['STATE'].str.contains('|'.join(search),na=False)]
west

#west.to_csv('C:/Users/MyPC/OneDrive/Documents/Inbiot VS/Zonal Distribution/WesternZone.csv')
